backend/requirements/base.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.5
pillow>=9.4.0
python-dotenv>=0.21.0
drf-yasg>=1.21.5
django-cors-headers>=3.13.0
python-decouple>=3.8.0
django-filter>=21.1
